Rename API names from json_*/JSON_* to pdjson_*/PDJSON_*
The json_* names are prone to clashing with other libraries and/or user
code. Plus, it makes sense to leave the clean json_* names to the application.

Note that we changed to pdjson_ rather that the uglier pdjson5_ because the
original pdjson project did not use this prefix and, no longer being
maintained, unlikely to ever do so.